Failing with zemu (#135)
* Failing with zemu;

* Working zemu test

* Clean up

* fixing clippy issues

* removing dependency

* restructure code
