+++ .circleci/config.yml
@@ -23,6 +23,7 @@ jobs:
     steps:
       - checkout
       - restore-sccache-cache
+      - run: sudo apt-get update -y && sudo apt-get install -y libusb-1.0.0 libudev-dev
       - run:
           name: rustfmt
           command: |
@@ -46,6 +47,7 @@ jobs:
     steps:
       - checkout
       - restore-sccache-cache
+      - run: sudo apt-get update -y && sudo apt-get install -y libusb-1.0.0 libudev-dev
       - run:
           name: rust unit tests --all-features
           command: |
