+++ examples/wasm_ledger/package.json
@@ -32,6 +32,7 @@
     "bip39": "^3.0.2"
   },
   "devDependencies": {
+    "@ledgerhq/logs": "^5.13.1",
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.9.0",
     "@babel/node": "^7.8.7",
