+++ examples/wasm_ledger/test/test.ledger.old.js
@@ -15,11 +15,33 @@
  ******************************************************************************* */
 const Zemu = require("@zondax/zemu").default;
 const Resolve = require("path").resolve;
+const LogListen = require("@ledgerhq/logs").listen
 
-const TransportNodeHid = require('@ledgerhq/hw-transport-node-hid').default;
+// const TransportNodeHid = require('@ledgerhq/hw-transport-node-hid').default;
 const signer_wasm = require('@zondax/filecoin-signer');
 
-describe("Ledger device", function () {
+const seed = "equip will roof matter pink blind book anxiety banner elbow sun young"
+const SIM_OPTIONS = {
+    logging: true,
+    start_delay: 4000,
+    X11: true,
+    custom: `-s "${seed}" --color LAGOON_BLUE`
+};
+
+before(async () => {
+    process.on("SIGINT", () => {
+        Zemu.stopAllEmuContainers(function () {
+            process.exit();
+        });
+    });
+    await Zemu.checkAndPullImage();
+})
+
+after(async () => {
+    await Zemu.stopAllEmuContainers();
+})
+
+describe.skip("Ledger device", function () {
   it("Get version", async function () {
       this.timeout(10000);
 
@@ -50,17 +72,33 @@ describe("Ledger device", function () {
   });
 })
 
-// Failing with Zemu
-it.skip("With Zemu", async function () {
+
+// FIXME: I am confused why this is the "legacy" version when it seems to be the newest one
+it("With Zemu", async function () {
     this.timeout(10000);
 
     const DEMO_APP_PATH = Resolve("bin/app.elf");
     const sim = new Zemu(DEMO_APP_PATH);
-    await sim.start({ logging: true });
+    await sim.start(SIM_OPTIONS);
 
     const transport = sim.getTransport()
-    console.log(transport.exchange)
-    var answer = await signer_wasm.getVersionFromDevice(transport);
 
+    // Subscribe to transport events to see what is going on...
+    LogListen( (s) => {
+        console.log(s);
+    } )
+
+    // FIXME:
+    // WASM is sending uint8array.. but HTTP transport expects Buffer.. so BOOM.. it breaks..
+    // We can instrument and convert from Buffer to uint8array but ideally WASM should send the correct type...
+    transport.old_exchange = transport.exchange
+    transport.exchange = async (apdu) => {
+        return transport.old_exchange( Buffer.from(apdu));
+    }
+
+    var answer = await signer_wasm.getVersion(transport);
     console.log(answer);
+    // FIXME: The reply is not correct. Ideally, we need to return the error code, etc
+    // We should be as backwards compatible as possible to what we already have
+
 });
