+++ ledger/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ledger-filecoin"
+name = "filecoin-signer-ledger"
 description = "Library to integrate with the Ledger Filecoin app"
 version = "0.10.0"
 license = "Apache-2.0"
@@ -13,11 +13,9 @@ edition     = "2018"
 autobenches = false
 
 [lib]
-name = "ledger_filecoin"
+name = "filecoin_signer_ledger"
 
 [dependencies]
-ledger-generic = { git = "https://github.com/Zondax/ledger-rs", branch = "feature/refactoring-wasm"}
-quick-error = "1.2.2"
 byteorder = "1.3.1"
 matches = "0.1.8"
 lazy_static = "1.2.0"
@@ -25,6 +23,11 @@ libsecp256k1 = "0.3.5"
 blake2b_simd = "0.5.10"
 futures = "0.3.4"
 serde = { version = "1.0", features = ["derive"] }
+thiserror = "1.0.16"
+
+# Used for internal development
+# ledger-transport = { path = "../../ledger-rs/ledger-transport"}
+ledger-transport = { git = "https://github.com/Zondax/ledger-rs", branch = "feature/refactoring-wasm"}
 
 # For wasm compiling
 [target.'cfg(target_arch = "wasm32")'.dependencies]
@@ -32,14 +35,9 @@ wasm-bindgen = "0.2.60"
 wasm-bindgen-futures = "0.4.10"
 js-sys = "0.3.37"
 
-# For not wasm compilation (native rust) compiling
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-ledger-transport = { git = "https://github.com/Zondax/ledger-rs", branch = "feature/refactoring-wasm"}
-
-
 
 [dev-dependencies]
-sha2 = "0.8.0"
+sha2 = "0.8.1"
 
 [profile.release]
 overflow-checks = true
