--- ledger/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "@ledgerhq/hw-transport-node-hid": "^5.13.1"
-  }
-}
