+++ ledger/transportWrapper.js
@@ -1,0 +1,6 @@
+
+export class TransportWrapper {
+    function exchange(apduCommand) {
+        return original_transport.exchange(apduCommand);
+    }
+}
